// FinalProjectUI/BookingInnUI/src/app/model/hotel.model.ts
import { Hotelfacilities } from "./hotelfacilities.model";
import { Location } from "./location.model";



export class Hotel{
    hid: number | undefined;
    hname: string | undefined;
    hdescription: string | undefined;
    haddress: string | undefined;
    hfetcher: string | undefined;
    hemail: string | undefined;
    hpassword: string | undefined;
    hphone: string | undefined;
    hwebsite: string | undefined;
    haverageprice: string | undefined;
    hcheckin: string | undefined;
    hcheckout: string | undefined;
    hpolicy: string | undefined;
    hrating: string | undefined;
    hlogo: string | undefined;
    hphoto: string | undefined;
    location: number | undefined;
    hotelfacilities: number | undefined;
        
}